requirements.txt
django==3.0.4
oauthlib==3.1.0
requests==2.23.0
requests_oauthlib==1.3.0
pyyaml==5.2
python-dateutil==2.8.1
